Added changes for change logo in ddr

+++ service-loans/src/main/java/com/capitaworld/service/loans/repository/fundprovider/ProductMasterRepository.java
@@ -85,6 +85,6 @@ public interface ProductMasterRepository extends JpaRepository<ProductMaster, Lo
 	public List<Integer> getProductsByOrgId(@Param("orgId")Long orgId);
 	
 	@Query("select user_org_id FROM fp_product_master pm where pm.id=:productId and pm.isActive = true")
-	public Long getUserOrgId(@Param("productId") Long productId);
+	public Integer getUserOrgId(@Param("productId") Integer productId);
 	
 }
